Add argument to control the auto select behaviour on accept_completion.
`wilder#accept_completion` can optionally take 0 or 1 as first argument
to control if the first element from selected directory gets completed
on the command line. The default value is 1 for consistency with the
current implementation.

The user facing `wilder#accept_completion` takes an optional argument
but the internal `wilder#main#accept_completion` makes the argument
mandatory.